refactoring suburbans
