+++ bot/handlers/suburbans/__init__.py
@@ -1,0 +1,1 @@
+from bot.handlers.suburbans import trails, personalization, updates
